Moved classes into more appropriate package
