+++ subprojects/logging/src/integTest/groovy/org/gradle/internal/logging/console/ConsoleDeprecationMessageTaskGroupingFunctionalTest.groovy
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.logging
+package org.gradle.internal.logging.console
 
 import static org.gradle.util.TextUtil.normaliseFileSeparators
